New Feature: Seeing Cummulative Voters Details
This Feature makes you see the total vote worth and vote count you received in a week from a user.

This feature also supports fiat conversion, so as to see vote worth in fiat currency.

You can also copy the details of this vote worth in steem-compatible format

This feature also allows you to view the list in Alphabetic Order, Order of Worth and Order of counts and all of this order have their own specific custom copy format